+++ sendmail.h
@@ -21,16 +21,14 @@ int wcSendMail(char *to, char *subj, char *body);
 
 struct buffer_st {
   char *data;
-  int length;
+  size_t length;
   char *ptr;
-  int offset;
+  size_t offset;
 };
 
 void buffer_new(struct buffer_st *b);
 void buffer_add(struct buffer_st *b, char c);
 void buffer_delete(struct buffer_st *b);
 
-void base64_encode(struct buffer_st *b, const char *source, int length);
-
-
+void base64_encode(struct buffer_st *b, const char *source, size_t length);
 #endif
